src/main/webapp/app/entities/client/update/client-update.component.ts
import { Component, OnInit } from '@angular/core';
import { HttpResponse } from '@angular/common/http';
import { ActivatedRoute } from '@angular/router';
import { Observable } from 'rxjs';
import { finalize, map } from 'rxjs/operators';

import { ClientFormService, ClientFormGroup } from './client-form.service';
import { IClient } from '../client.model';
import { ClientService } from '../service/client.service';
import { IUser } from 'app/entities/user/user.model';
import { UserService } from 'app/entities/user/user.service';
import { IResourceData } from 'app/entities/resource-data/resource-data.model';
import { ResourceDataService } from 'app/entities/resource-data/service/resource-data.service';

@Component({
  selector: 'jhi-client-update',
  templateUrl: './client-update.component.html',
})
export class ClientUpdateComponent implements OnInit {
  isSaving = false;
  client: IClient | null = null;

  usersSharedCollection: IUser[] = [];
  resourceDataSharedCollection: IResourceData[] = [];

  editForm: ClientFormGroup = this.clientFormService.createClientFormGroup();

  constructor(
    protected clientService: ClientService,
    protected clientFormService: ClientFormService,
    protected userService: UserService,
    protected resourceDataService: ResourceDataService,
    protected activatedRoute: ActivatedRoute
  ) {}

  compareUser = (o1: IUser | null, o2: IUser | null): boolean => this.userService.compareUser(o1, o2);

  compareResourceData = (o1: IResourceData | null, o2: IResourceData | null): boolean =>
    this.resourceDataService.compareResourceData(o1, o2);

  ngOnInit(): void {
    this.activatedRoute.data.subscribe(({ client }) => {
      this.client = client;
      if (client) {
        this.updateForm(client);
      }

      this.loadRelationshipsOptions();
    });
  }

  previousState(): void {
    window.history.back();
  }

  save(): void {
    this.isSaving = true;
    const client = this.clientFormService.getClient(this.editForm);
    if (client.id !== null) {
      this.subscribeToSaveResponse(this.clientService.update(client));
    } else {
      this.subscribeToSaveResponse(this.clientService.create(client));
    }
  }

  protected subscribeToSaveResponse(result: Observable<HttpResponse<IClient>>): void {
    result.pipe(finalize(() => this.onSaveFinalize())).subscribe({
      next: () => this.onSaveSuccess(),
      error: () => this.onSaveError(),
    });
  }

  protected onSaveSuccess(): void {
    this.previousState();
  }

  protected onSaveError(): void {
    // Api for inheritance.
  }

  protected onSaveFinalize(): void {
    this.isSaving = false;
  }

  protected updateForm(client: IClient): void {
    this.client = client;
    this.clientFormService.resetForm(this.editForm, client);

    this.usersSharedCollection = this.userService.addUserToCollectionIfMissing<IUser>(this.usersSharedCollection, client.user);
    this.resourceDataSharedCollection = this.resourceDataService.addResourceDataToCollectionIfMissing<IResourceData>(
      this.resourceDataSharedCollection,
      client.company
    );
  }

  protected loadRelationshipsOptions(): void {
    this.userService
      .query()
      .pipe(map((res: HttpResponse<IUser[]>) => res.body ?? []))
      .pipe(map((users: IUser[]) => this.userService.addUserToCollectionIfMissing<IUser>(users, this.client?.user)))
      .subscribe((users: IUser[]) => (this.usersSharedCollection = users));

    this.resourceDataService
      .query()
      .pipe(map((res: HttpResponse<IResourceData[]>) => res.body ?? []))
      .pipe(
        map((resourceData: IResourceData[]) =>
          this.resourceDataService.addResourceDataToCollectionIfMissing<IResourceData>(resourceData, this.client?.company)
        )
      )
      .subscribe((resourceData: IResourceData[]) => (this.resourceDataSharedCollection = resourceData));
  }
}
